all: move utility packages under a 'util' package
