+++ .github/workflows/release.yml
@@ -17,8 +17,8 @@ env:
   GO_VERSION: "1.24.x"
   PNPM_VERSION: "9.4.x"
   GO_LDFLAGS: >-
-    -X 'github.com/hemilabs/heminetwork/version.Brand=Hemi Labs'
-    -X github.com/hemilabs/heminetwork/version.PreRelease=
+    -X 'github.com/hemilabs/heminetwork/util/version.Brand=Hemi Labs'
+    -X github.com/hemilabs/heminetwork/util/version.PreRelease=
 
 jobs:
   # Build and test
