+++ bitcoin/wallet/gozer/tbcgozer/tbcgozer_test.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/hemilabs/heminetwork/bitcoin/wallet/gozer"
 	"github.com/hemilabs/heminetwork/service/tbc"
-	"github.com/hemilabs/heminetwork/testutil"
+	"github.com/hemilabs/heminetwork/util/testutil"
 )
 
 func TestTBCGozer(t *testing.T) {
