+++ cmd/hproxyd/hproxyd.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/hemilabs/heminetwork/config"
 	"github.com/hemilabs/heminetwork/service/hproxy"
-	"github.com/hemilabs/heminetwork/version"
+	"github.com/hemilabs/heminetwork/util/version"
 )
 
 const (
