+++ cmd/keygen/keygen.go
@@ -17,7 +17,7 @@ import (
 	dcrsecpk256k1 "github.com/decred/dcrd/dcrec/secp256k1/v4"
 
 	"github.com/hemilabs/heminetwork/ethereum"
-	"github.com/hemilabs/heminetwork/version"
+	"github.com/hemilabs/heminetwork/util/version"
 )
 
 var (
