+++ database/tbcd/level/encodedecode_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/hemilabs/heminetwork/database/tbcd"
 	"github.com/hemilabs/heminetwork/hemi"
-	"github.com/hemilabs/heminetwork/testutil"
+	"github.com/hemilabs/heminetwork/util/testutil"
 )
 
 func TestKeystoneEncodeDecode(t *testing.T) {
